bridge/requirements.txt
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
pystray>=0.19.0
Pillow>=10.0.0
cryptography>=42.0.0
pyinstaller>=6.0.0
pytest>=7.0.0
